vista de dentistas completa

+++ web-consultorio/consultorio-odontologico/src/components/DentistList.jsx
@@ -1,6 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
-import FormUpdate from '../components/FormUpdate';
+import FormUpdateDentist from './FormUpdateDentist';
+import FormCreateDentist from './FormCreateDentist';
+import Swal from "sweetalert2"
 
 
 
@@ -8,13 +10,16 @@ import FormUpdate from '../components/FormUpdate';
 export default function DentistList() {
   const [list, setList] = useState([]);
   const [updating, setUpdating] = useState([false, -1]);
-  const [dentistToUpdate, setDentistToUpdate] = useState();
+  const [creating, setCreating] = useState(false);
+
 
 
   useEffect(() => {
     //funciona como componentDidMount
     updateList()
   }, [])
+
+
 
   const updateList = () => {
     let requestOptions = {
@@ -32,26 +37,56 @@ export default function DentistList() {
   }
 
   const deleteDentist = (id) => {
+
     console.log(id)
 
-    let requestOptions = {
-      method: 'DELETE',
-    };
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Deleted!',
+          'El dentista fue borrado.',
+          'success'
+        )
 
-    fetch("http://localhost:8080/api/dentist/" + id, requestOptions)
-      .then(response => response.text())
-      .then(result => {
-        updateList()
-        return console.log(result)
-      })
-      .catch(error => console.log('error', error));
+        let requestOptions = {
+          method: 'DELETE',
+        };
+
+        fetch("http://localhost:8080/api/dentist/" + id, requestOptions)
+          .then(response => response.text())
+          .then(result => {
+            updateList()
+            return console.log(result)
+          })
+          .catch(error => console.log('error', error));
+      }
+    })
+
+
+  }
+
+  const cancelCreate = () => {
+    setCreating(false)
   }
 
   const cancelUpdate = () => {
     setUpdating([false, -1])
   }
 
-  const modifyDentist = (id) => {
+  const createDentistDisplay = () => {
+    setCreating(true)
+  }
+
+  const modifyDentistDiplay = (id) => {
     console.log(id)
     setUpdating([true, id])
 
@@ -60,6 +95,7 @@ export default function DentistList() {
 
   return (
     <div>
+      <button onClick={() => createDentistDisplay()} className='btn btn-primary' style={{ margin: "20px" }}>Create Dentist</button>
       <table className="table mx-auto" style={{ maxWidth: "80%" }}>
         <thead>
           <tr>
@@ -75,14 +111,15 @@ export default function DentistList() {
                 <th scope="row">{ele.id}</th>
                 <td key={"2" + ele.name}>{ele.name}</td>
                 <td key={"3" + ele.lastName}>{ele.lastName}</td>
-                <td onClick={() => modifyDentist(ele.id)} className='btn btn-outline-success'> Modify </td>
+                <td onClick={() => modifyDentistDiplay(ele.id)} className='btn btn-outline-success'> Modify </td>
                 <td onClick={() => deleteDentist(ele.id)} className='btn btn-outline-danger delete-button'> X </td>
               </tr>
             )
           })}
         </tbody>
       </table>
-      {updating[0] ? <FormUpdate target={updating[1]} update={updateList} cancelUpdate={cancelUpdate} /> : ""}
+      {creating ? <FormCreateDentist cancelCreate={cancelCreate} update={updateList} /> : ""}
+      {updating[0] ? <FormUpdateDentist target={updating[1]} update={updateList} cancelUpdate={cancelUpdate} /> : ""}
 
     </div>
   )
